separate file downloader and main
